AutoCoder-Forge-Basic/requirements.txt
jinja2
pyyaml
watchdog
black
pytest
fastapi
uvicorn[standard]
pygame
pyinstaller
